# speedup hashing function by doing less mixing