I tested with ./benchmark  --benchmark_repetitions=100 --benchmark_report_aggregates_only=true

old code

BM_CreateScopeLinkWithVariableList_mean            1558 ns       1558 ns     452391
BM_CreateScopeLinkWithVariableList_median          1551 ns       1551 ns     452391
BM_CreateScopeLinkWithVariableList_stddev            29 ns         29 ns     452391
BM_CreateScopeLinkWithoutVariableList_mean         1169 ns       1169 ns     603959
BM_CreateScopeLinkWithoutVariableList_median       1163 ns       1163 ns     603959
BM_CreateScopeLinkWithoutVariableList_stddev         24 ns         24 ns     603959
BM_CreateScopeLinkWithLambdaLink_mean              1065 ns       1065 ns     638571
BM_CreateScopeLinkWithLambdaLink_median            1059 ns       1059 ns     638571
BM_CreateScopeLinkWithLambdaLink_stddev              23 ns         23 ns     638571


new

BM_CreateScopeLinkWithVariableList_mean            1548 ns       1547 ns     452378
BM_CreateScopeLinkWithVariableList_median          1531 ns       1531 ns     452378
BM_CreateScopeLinkWithVariableList_stddev            36 ns         36 ns     452378
BM_CreateScopeLinkWithoutVariableList_mean         1158 ns       1158 ns     605659
BM_CreateScopeLinkWithoutVariableList_median       1152 ns       1152 ns     605659
BM_CreateScopeLinkWithoutVariableList_stddev         20 ns         20 ns     605659
BM_CreateScopeLinkWithLambdaLink_mean              1054 ns       1054 ns     615944
BM_CreateScopeLinkWithLambdaLink_median            1049 ns       1049 ns     615944
BM_CreateScopeLinkWithLambdaLink_stddev              21 ns         21 ns     615944