@vsbogd 2018-06-22
@noskill, are you sure this optimization changes something for Release build mode? I looked at master release version of ScopeLink::compute_hash and it seems that compiler do the same optimization with fnv1a_hash() inlined. May be it make sense implement pure fnv1a_hash() microbenchmark to show the difference?