@linas 2018-06-22
>  compiler do the same optimization

The old code ran the loop one byte at a time.  The new code runs it 8 bytes at a time.  It is possible that the compiler was able to loop-unroll the old code, and optimize it using the amd64 byte-parallel instructions to run 8 at a time. This would not surprise me.